photo-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# Supporting packages
python-dotenv>=0.19.0  # For environment variables
pyOpenSSL>=20.0.0      # For secure connections
requests